Show Page: Quick Action Modal
I decided to expand the width of the buttons to the full width of the modal to improve accessibility.

Doing this may cause a hindrance in accessibility of the "x" mark button inside the modal.

Since the modal is automatically closed when not "focused" or "hovered" on, I have decided to remove the "x" button.