Seearch Engine
Made a massive revamp to the functionality of the search engine.

In the old search engine, there was a massive bug involving the appending of results into the DOM.

The first result is displayed, the the corresponding result is appended into the structure of the preceding result.
This process continued for all results that where displayed.

Though there was no physical difference in the appearance, this problem could become disastrous in future design modifications.

With the new design, UI changes will need to be made. I am also remove the search bar from the navbar and have all searches done from the search page itself