Episode Page: More Ep Modal
I have ccorrently halted the application of this modal. Implementing it in it's current state will prove vert=y tedius in the event of shows with great number of episodes.

I plan to formulate a better way to implement this modal.

In the meantime clicking the button sends you to episode section of the main page.