Adding New Profiles
I attempted to functionalize this as users may want to create a new profile. It partially works but one issue.

All newly added "profile cards" do not respond to the event listeners declared before their insertion.

I am still working on this but for now, if a user clicks the  "Add profile" button an error notification is displayed. I found that better than not showing anything.