Navbar Side(hotdog) Menu
I updated the error message for highlighting navbar items corresponding to the current page.
The old message: "error" was not descriptive but instead confusing, even to me the developer.
I have changed it to: "Error: No corresponding navbar item found".
This clearly states there is an error and briefly describes what that error is.