Show Pages
I am testing a new way to dynamically load the content of the show pages.
Doing this will reduce the amount of time spent hardcoding each content for individual shows.

To achieve this, I have created a central file to store all the information for individual shows. The upon loading of the shows, the corresponding information is retrieved and used to update the content of the pages.