Updates
1. Navbar Side (hotdog) Menu:
I replaced the "My List" with the three categories that it's comprised of.
I also renamed "Popular" to "Trending" and have removed the "Recent" link as I am undecided as to whether to include it or not.

2. Trending Page:
I have started developing the Trending page.

3. Section Pages:
I removed the size limiter for the basic slide cards.