Basic Slider Carousel Scrolling
I updated the scrolling of the carousel to only scroll by one card instead of the client width.
This should make it easier for users to scroll through the carousel's content.

I also added a new condition that removes/adds the "hide" class to the carousel's arrows if there is no overflow.