Navbar Profile Menu
I noticed that on mobile browsers, the size of the app's action bar is not a part of the viewport. Thus, when dragged into view, it pushes the profile menu out of view.

Unfortunately, the height differs across browsers and no fixed value can be used. To combat this, I updated my calculation to measure only 90% of the viewport height instead of 100%.