Calibrating Navbar Notification Menu
There was an issue with the previous calculations. If the viewport height is lower than the menu's height during the initial loading, the menu's height becomes fixed and fails to extend when the device's height becomes larger.

To battle this, rather than recording the height of the menu's parent container, I measure the heights of the sub-sections, then add them up to set the menu's height. This fixed the issue.