Navbar and User Notifications
I noticed some bugs that only occurred when the user is in the 'Notifications' page.

1. Since the notification page fetches notifications, the library gets called twice. I fixed this by adding a conditional statement which checks whether the script is in the DOM by targeting its 'id' attribute.
If it is undefined, then the script will be created and inserted. If defined, then the fetching function is called instantaneously.

2. The 'Mark al as read' button also caused certain functions to be called twice as they were similar. I fixed this by having separate listeners for both buttons.
I then modified the 'user notifications' button's click event to simulate a click event on the 'navbar notifications' button.