MAJOR UPDATE
After many months of debugging, I found the cause behind the excessive whitespace between cards in my grid. it was due to a typo. i.e. I typed "auto-fit" instead of "auto-fill".
This caused the columns to stretch to fit the number of cards. This might not be an issue on mobile devices, but it posed a nightmare on PCs and desktops.
I am super glad to have finally fixed this😪