User pages
I added a new feature that updates the welcome message and the background image of the header box whenever the user switches profiles.
This means that the content of the welcome message will not contain the user name but rather the current profile name.
The username can be seen in the navbar profile dropdown menu and the user profile page.