File Path Changes | Anime -> TV
After a long delay, I changed the file path, from 'Anime' to 'TV'.  This change was supposed to have been effected after updating the Show categories from 'Movies', 'TV', 'Cartoon', and 'Anime' to just 'TV' and 'Movies'.

This change was made earlier because 'Cartoon' and 'Anime'  had both series and movies and thus couldn't be a section on their own.

However, I delayed changing the file path for several reasons. This change became mandatory as I work on properly displaying information for movies and series.