User - Interface Update
I have revamped the layout of the app, specifically focusing on the navbar and icons. The new navbar layout  makes navigating easier with a nice and comfy look. I was able to achieve this while using a lower footprint.

I also updated the logo of the app and icons. I will be implementing this change globally.