Compatibility issue
I noticed on FIrefox, the href tag behaves differently and assumes an inline display if not specified. This prevents it from reading dimensions (in percentage for my case).
The end result was some misaligned content.

To fix this, I used "divs" to wrap the nested elements then placed the href tag as an absolute element.