Quick Search Shortcut
To close then navbar menus, a click event had to be registered. However the quick search shortcut is triggered by the key-down event. This means that when opened, the navbar menus remain open, causing inconveniences.

To fix this, I made the open-quick-search element a button and the shortcut will simulate a click of the button rather than calling the function directly.