Selector Accessibility | Schedule Page
I fixed a huge accessibility issue with the selectors (dates and filters).

Previously, the selectors sit at the top of the page and if the user scrolls past them while viewing content, the only way to access them was to scroll to the top of the page. This would be really bad in large content and a nightmare for mobile users.

This fix should make accessing the selectors seamless across different platforms.