I updated "sticky" selector bar on large screens
The filter is displayed instead of the date selector when "sticky". In the previous version, hovering displays the filter selector which increases bar's height and pushes down the date selector away from the cursor.

By switching the selectors displayed, the cursor is placed directly over the date selector thus resolving that issue. However, users cannot see the current date for the content they're viewing unless they hover the bar.

To fix this, I added a small display containing the current date of the content being viewed. This element will only be displayed when the date selector is not visible.