Update schedule.js
I found out the issue occurred because the transition ended before the date slider was displayed. Therefor executing the scroll function at that time didn't have any effect.

At the moment, I am using a timeout function to perform this scroll-event. However, in the event of large content, this approach may be unreliable.