Hide/Unhide navbar while scrolling
I noticed that if user triggers full-screen from the browser, the document cannot register that as full-screen. This was fixed by comparing the screen height to the browser height.

However, on mobile browsers, full-screen only made the window and screen height equal in landscape mode. And the condition first checks this comparism which always results in false as they are not equal on portrait.

To resolve this i switched the positions of the conditions and changed "OR" to "AND". With this switch, the condition checks any element for full-screen first and must satisfy BOTH instead of one to work.