Gen Menu Modal
The previous menus had many issues, the primary one being constant repetition and re-declaration of variables.

The old solution to this only included a partial number of menus and was fixed to the 'User Pages'. This means that the functions couldn't be used globally and only within the watchlist page.

Therefore, I developed a general menu modal that will provide access to all current and future menus for all pages in the app.

Although the mobile layout was successfully implemented, the pc version is still undergoing tests.

I am also revising certain functions to work properly with the menu.