Mapping the Search Inventory | Watchlist Page
I noticed the previous map declaration ran into issues because the search inventory hadn't finished loading by the time the declaration was made.

This broke certain parts of the page forcing the user to reload to fix it.

I revised it so that the declaration is called only after the inventory has finished loading.