WL Modal | Watchlist Page
If the modal is closed and re-opened very quickly, the content fails to load. This occurred because, the condition to remove the previous content was based on an independent timer whose value was less than the transition time of the modal.

To fix this, the previous content is removed immediately the transition of the modal ends.