Schedule Page Update (Majorly mobile)
Since, the page is automatically scrolled back to the top after switching dates, the date selector containers will simply remain fixed to the top of the screen as in the case of pc screens.

However as hovering isn't a standardly accessible feature of mobile, the content type selector will be hidden as the user scrolls. I initially considered just displaying the box while scrolling but their current size in addition to the fixed bottom nav bar may take too much screen space.