New List: Collections
Previously, the user's watchlist was essentially a fixed number of playlists where shows can be added to. It was a failing attempt to combine two distinct functionalities into one feature.

With the addition of "Collections", users can now create and customize their own unique playlists. This is separate from the standard watchlist and is unique to every profile.