Home page Revamp
For the longest time, users have been stuck with the bland lifeless home page, that looked like a premature Crunchyroll homepage (The inspiration for its design).

It has finally gotten a much needed update embodying a new fresh look with randomized sections to introduce a sense of dynamicity and improve overall accessibility.