Default Action for Navigation (Sign In users)
I updated the action taken when user types in a wrong hash or no hash for the page navigation. Previously, it displayed the default page but will now redirect to the homepage.

This is for signed in user navigation. Users not signed in will still be redirected to the landing page.