Select Profile Pic Carousels
I figured out the problem to be an indexing issue. Basically, since the frg slider was initialized before the bcg slider, when ever the basic slider is initialized, it always indexes the frg first.
This means that the Bounding Client Rect composition was determined by the frg. If the frg has display none, its always zero, and reverse is the case when frg is visible.

To fix this, ONLY the sel pic slider in use is initialized while the other is not (or removed) and is removed completely after use.