Show Pages | Background Image Break/Overlap
Although the "ImgHeight" property is automatically calculated and updated, it is done after the site has been initialized. Meaning, it is undefined during parsing. This causes content to partially break through (overlap)  the other.

To solve this, I will define an initial value for the property before the automated logic fires.