Navigation Routing System (Beta)
Yes, I should take a seat and learn React, Angular, or some other one and use their already polished and fine-tuned routing system which I will...soon.

At the moment, I have made some updates to my routing system.
1. I changed "Uvid" to "Uvid+"

2. I moved the app name from the front of the document title to the end of it. This should allow the users see the specific page they are on, before they see the app name (which they may already know).

3. I have replaced the "join" route with two separate routes:
-> "signin" = Handles user sign in logic and password reset logic.
-> "signup" = Handles the creation and setup of the user's account.

4. I updated the route logic to correspond with the Firebase functions. It works however, there are bugs or issues that occur.

5. Speaking of bugs, the "defer_page_route" function which allowed user to return to the page they attempted to access while logged out is currently placed on pause. Unlike most functionalities, it hasn't adapted well to Firebase's presence.
It's an extra feature and not a fundamental on, thus it's absence is not coherently disadvantageous. However, I believe it's a good feature to have, and will continue to work on it.