Quick Search is Back!
I've worked on and hopefully, fixed all the old bugs brought by the old quick search. With that said, I'll be adding it back to Uvid.

It's shortcut has been updated to "Ctrl" + "/" as it's a more common (and easier/quicker) key combination than the previous.

I also updated the icons for both pc and mobile navbars.