Help Page | Contact
I patched a bug with the contact chatbot (which doesn't work) that allowed the user to perform XSS.

I also switched the "submit" and "new-line" keyboard shortcuts with each other to prevent confusion. Although, the "Enter" key is normally used for submission, it's used for new-lines in <textarea> fields. So it felt right to leave it that way. This was especially needed on mobile devices which don't provide the "Shift" key by default.

Shift + Enter ==> send
Enter ====> new-line