Update main.js
I noticed an interesting bug. If the user is logged in on two devices at the same time, then selects a different profile and customizes it, the previously selected profile's properties gets overwritten by those changes.

I believe the issue births from the logic that determines whether to update the selected profile, or a different profile based on its id.

I am testing a possible fix for it.