Font Style Redesign - Zen Dots...it's time
While the Zen Dots was horrid for reading as found in v2 of this project, pairing it with a almost universally readable font; Poppins, seemed to do the trick.

However, the issue of readability, particularly on smaller screens (mobile phones) was still a issue. After much debate, I've decided to switch back to the "Berlin Sans FB" font, originally used in the 1st version.

I've made this decision primarily for readability and uniformity. Nonetheless, I'm honestly still conflicted on this change. I think the Zen Dots + Poppins paired well but I wanted a single font family, and neither of the two were good enough on their own.

I am effecting this across the site. The new font is somewhat smaller than the previous by default. Not sure why but it might originate from the their respective designers.