Search Result Display Improvement
I updated the logic for displaying the results from searches. The old method interacted with the DOM directly as it built and updated results. This sometimes created instances where the layout shifts or breaks as space is created for a newly built result to be inserted.

The new way ensures that all results are built up before insertion, thereby eliminating layout shifts/breaks and slightly improving performance.